% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCimplify_for_Seurat.R
\name{SCimplify_for_Seurat}
\alias{SCimplify_for_Seurat}
\title{SCimplify_for_Seurat}
\usage{
SCimplify_for_Seurat(
  seurat,
  seurat.mc = NULL,
  k.knn = 30,
  kith = NULL,
  kernel = T,
  gamma = 20,
  graph.name = NULL,
  assay = c("RNA"),
  reduction = list("pca"),
  dims = list(c(1:30)),
  membership = NULL,
  metacellNormalization = F,
  avg.in.data = F,
  fragmentFiles = NULL,
  tmpPath = NULL,
  outputDirMcFragment = NULL,
  bgzip_path = NULL,
  tabix_path = NULL,
  prefixMC = "",
  seed = 12345L,
  return_membership_table = TRUE,
  return_fragment_manifest = TRUE,
  peakSep = c("-", "-"),
  label = NULL,
  return.seurat = T,
  nb_cl = NULL,
  verbose = FALSE
)
}
\arguments{
\item{seurat}{A Seurat single-cell object. It has to be preprocessed (eg. latent space computed) for the assay(s) used to identify metacells}

\item{gamma}{graining level.}

\item{assay}{a list of one or two assays to use to build the knn graph on which metacell are identified.}

\item{reduction}{a list of corresponding reduction name in the seurat single cell object.}

\item{dims}{a list of corresponding dimensions to use.}

\item{membership}{a vector of metacell membership as in SuperCell v1 to use to directly aggregate the data (optionnal).}

\item{sobj.mc}{A metacell seurat object that will be rescaled (optional) it requires a metacell_hierarchy object in the slot misc.}

\item{return_fragment_manifest}{whether to store fragment_manifest and fragment_cell_map tables in misc.}

\item{seed}{integer seed used for reproducible metacell construction.}

\item{return_membership_table}{whether to store the membership table in misc.}
}
\value{
A Seurat metacell object with metacell_hierarchy and memberships in the slot misc.
}
\details{
\code{sample_col} is not part of the SuperCell2 builder API. To build
sample-specific metacells, split the input Seurat object by its sample
metadata before calling this function. Use \code{label} only for optional
semi-supervised metacell construction.
}
\description{
\code{SCimplify_for_Seurat} 
Build metacells from a Seurat single-cell object.
}
\examples{
sobj.mc <- SCimplify_for_Seurat(seurat = pbmc,
                         gamma = 30)
}
