% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_fragment_file.R
\name{AggregateFragmentFile}
\alias{AggregateFragmentFile}
\title{AggregateFragmentFile parallel}
\usage{
AggregateFragmentFile(
  input_file,
  membership,
  output_name = NULL,
  output_path = NULL,
  tmp_path = "./tmp/",
  nb_cl = NULL,
  nb_row_split = 10000000L,
  bgzip_path = NULL,
  tabix_path = NULL,
  returnOutputFileName = TRUE,
  keep_unmatched = FALSE
)
}
\arguments{
\item{input_file}{path to original fragments file}

\item{membership}{named vector mapping single-cell barcodes to final metacell barcodes}

\item{output_name}{name of the metacell fragments file}

\item{output_path}{directory where the new fragments file will be saved}

\item{tmp_path}{directory path for temporary files}

\item{nb_cl}{number of workers to use}

\item{nb_row_split}{row number for split files}

\item{bgzip_path}{optional path to bgzip}

\item{tabix_path}{optional path to tabix}

\item{returnOutputFileName}{whether to return the new fragment file name}

\item{keep_unmatched}{whether to keep fragments whose barcode is absent from membership}
}
\description{
Transform fragments from single-cell barcodes to metacell barcodes.
}
}
