% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetacellExpression.R
\name{MetacellExpression}
\alias{MetacellExpression}
\title{MetacellExpression}
\usage{
MetacellExpression(
  object,
  pb.method = "aggregate",
  assays = NULL,
  features = NULL,
  return.seurat = TRUE,
  group.by = "ident",
  add.ident = NULL,
  slot = "counts",
  layer = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat single-cell object. It has to be preprocessed (eg. latent space computed) for the assay(s) used to identify metacells}

\item{slot}{Assay slot/layer to aggregate. Use `layer` as an alias for Seurat v5 callers.}

\item{layer}{Deprecated alias of `slot`; retained for Seurat v5 compatibility.}
}
\value{
A Seurat object with aggregated data
}
\description{
\code{MetacellExpression} 
Compute metacells from a Seurat single-cell object.
}
