% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCimplify_for_Seurat.R
\name{SCimplify_for_Seurat}
\alias{SCimplify_for_Seurat}
\title{SCimplify_for_Seurat}
\usage{
SCimplify_for_Seurat(
  seurat,
  seurat.mc = NULL,
  k.knn = 30,
  kith = NULL,
  kernel = T,
  gamma = 30,
  graph.name = NULL,
  assay = c("RNA"),
  reduction = list("pca"),
  dims = list(c(1:30)),
  membership = NULL,
  metacellNormalization = F,
  avg.in.data = F,
  fragmentFiles = NULL,
  tmpPath = NULL,
  outputDirMcFragment = NULL,
  bgzip_path = NULL,
  tabix_path = NULL,
  prefixMC = "",
  seed = 12345L,
  return_membership_table = TRUE,
  return_fragment_manifest = TRUE,
  peakSep = c("-", "-"),
  label = NULL,
  return.seurat = T,
  nb_cl = NULL,
  verbose = FALSE
)
}
\arguments{
\item{seurat}{A Seurat single-cell object. It has to be preprocessed (eg. latent space computed) for the assay(s) used to identify metacells.}

\item{seurat.mc}{An optional metacell Seurat object to rescale. It must contain a metacell hierarchy in \code{misc}.}

\item{gamma}{Graining level. The default is 30 cells per metacell.}

\item{assay}{A list of one or two assays used to build the graph on which metacells are identified.}

\item{reduction}{A list of corresponding reduction names in the single-cell Seurat object.}

\item{dims}{A list of corresponding dimensions to use.}

\item{membership}{An optional vector of existing metacell memberships used to aggregate the data directly.}

\item{label}{Optional metadata column used to keep labeled groups separate during metacell construction. This may contain condition, sample, cell type, or another categorical annotation; \code{NA} values enable partial annotation.}

\item{return_fragment_manifest}{Whether to store fragment-manifest and fragment-cell-map tables in \code{misc}.}

\item{seed}{Integer seed used for reproducible metacell construction.}

\item{return_membership_table}{Whether to store the membership table in \code{misc}.}
}
\value{
A Seurat metacell object with the metacell hierarchy and memberships in \code{misc}.
}
\details{
\code{sample_col} and \code{condition_col} are not formal arguments of this Seurat builder. Pass a Seurat metadata column name through \code{label} when metacells must not mix known conditions, samples, cell types, or other labeled groups. The matrix builders \code{SCimplify()} and \code{SCimplify_from_embedding()} instead accept condition and cell-type vectors through \code{cell.split.condition} and \code{cell.annotation}, respectively.
}
\description{
Build metacells from a preprocessed Seurat single-cell object.
}
\examples{
sobj.mc <- SCimplify_for_Seurat(seurat = pbmc, gamma = 30)
}
